Make item extend document, and bring its metadata all the way to the composition space! Use conditionals to get the best field from item into its location field.
Make semantic actions & getContainer() methods pass metadata through properly when possible.
This was a pain in the present neanderthal container architecture, but it seems to work.

Was forced to promote MetaMetadataCompositeField as an alternative to MetaMetadata in a number of places.
Kind of a pain.
<<Need to discuss with Andrew & rest of team.>>

More clean-up of semantic action code. Use new style more convenient argument value accessors.

Add null check semantic action.