metadata expansion further implementation: looking up for downloaded documents during extraction (for xpath only right now).
tracking download status in Document (should merely be a clone of DocumentClosure.downloadStatus; we need it in Document because it needs to be stored in database).

CompositeTermVector fix: do bookkeeping during notifyObservers() and update(), so that one observer will not be updated twice or more. this prevents infinite loop on notifyObservers() and update().