Commit - Generated Semantics use generics
Problem solved:

 Some metadata that extend Search, such as NsfSearch, specify the generic as follows:

<meta_metadata name="nsf_search" extends="search" parser="xpath">
<generic_type_var name="SR" arg="nsf_award" />

others do not, and therein lies the problem. the c# code generator needs to specify the generic using the genericTypeVars field in meta-metadata. this field should be currently inherited to meta-metadata that extends a base class with generics. the code generator needs to check for this genericTypeVars. if it exists and no arg attribute is defined, then it should use the extends attribute to specify the generic.

for example, the meta-metadata for delicious_search which extends search does not specify a generic_type_var. this currently generates the following code, which does not compile:

public class DeliciousSearch : Search

it should generate the follow code:

public class DeliciousSearch : Search<Document>

the document generic specification is coming from the extends attribute for the generic_type_var in the search meta-metadata:

<meta_metadata name="search" extends="document" comment="">
	  <generic_type_var name="SR" extends="document" />